refactored to use 2 containers, one for routing, one for logic
