+++ src/components/ApplicationDashboard.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React from 'react'
 import styled from 'styled-components'
 import { H1 } from './Typography'
 import { Button } from './Input/Button'
@@ -8,10 +8,6 @@ import instagram from '../assets/icons/instagram.svg'
 import medium from '../assets/icons/medium.svg'
 import twitter from '../assets/icons/twitter.svg'
 import { ReactComponent as HandWave } from '../assets/hand-wave.svg'
-import { useAuth } from '../utility/Auth'
-import { useHackerApplication } from '../utility/HackerApplicationContext'
-import { useLocation } from 'wouter'
-import { getLivesiteDoc } from '../utility/firebase'
 
 const Container = styled.div`
   margin: 5em auto;
@@ -136,38 +132,26 @@ const SocialMediaLinks = () => {
   )
 }
 
-const Dashboard = ({ hackerStatus }) => {
-  const { user } = useAuth()
-  const { updateApplication, forceSave } = useHackerApplication()
-  const [, setLocation] = useLocation()
-  const [livesiteDoc, setLivesiteDoc] = useState(false)
-  const canRSVP =
-    hackerStatus === 'acceptedNoResponseYet' || hackerStatus === 'acceptedNotAttending'
-  const setRSVP = canRSVP => {
-    updateApplication({
-      status: {
-        responded: true,
-        attending: canRSVP,
-      },
-    })
-    forceSave()
-  }
-  useEffect(() => {
-    const unsubscribe = getLivesiteDoc(setLivesiteDoc)
-    return unsubscribe
-  }, [setLivesiteDoc])
+const Dashboard = ({
+  hackerStatus,
+  isApplicationOpen,
+  canRSVP,
+  setRSVP,
+  username,
+  editApplication,
+}) => {
   return (
     <Container>
       <WelcomeHeader>
-        <WelcomeMessage>Welcome Back, {user.displayName}</WelcomeMessage>
+        <WelcomeMessage>Welcome Back, {username}</WelcomeMessage>
         <StyledHandWave />
       </WelcomeHeader>
       <AppLinks>
         <HackerAppText>YOUR HACKER APPLICATION</HackerAppText>
         <EditAppButton
           color="secondary"
-          onClick={livesiteDoc.applicationsOpen && (() => setLocation('/application/part-1'))}
-          disabled={!livesiteDoc.applicationsOpen}
+          onClick={isApplicationOpen && (() => editApplication())}
+          disabled={!isApplicationOpen}
         >
           Edit Your Application
         </EditAppButton>
@@ -195,10 +179,10 @@ const Dashboard = ({ hackerStatus }) => {
           <SocialMediaLinks />
           <RSVPButton
             width="flex"
-            onClick={livesiteDoc.applicationsOpen && (() => setRSVP(canRSVP))}
+            onClick={isApplicationOpen && (() => setRSVP(canRSVP))}
             shouldDisplay={canRSVP || hackerStatus === 'acceptedAndAttending'}
             color={canRSVP ? 'primary' : 'secondary'}
-            disabled={!livesiteDoc.applicationsOpen}
+            disabled={!isApplicationOpen}
           >
             {canRSVP ? 'RSVP' : 'un-RSVP'}
           </RSVPButton>
