+++ src/pages/Application/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Dashboard from '../../components/ApplicationDashboard'
+import Dashboard from '../../containers/Application/Dashboard'
 
 // applicant dashboard
 export default ({ hackerStatus }) => {
